third_party/blink/web_tests/external/wpt/css/css-properties-values-api/animation/custom-property-transition-transform-list-box-size.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] A custom property of type <transform-list> yields a CSS Transition using the mix function for a box size dependent matrix interpolation
  assert_equals: Element has the expected animated value expected "mix(50%; translateX(50%); scale(4))" but got "scale(4)"
Harness: the test ran to completion.
